backend/requirements.txt
fastapi
uvicorn
pydantic
httpx
xarray
netcdf4
numpy
python-dotenv
requests